Network proto handlers/container fixes (#6334)
* Fix HP transport + some double <-> float problems

TOCLIENT_HP transport u16 hp as a u8, use u16 HP, this prevent HP over 255 to overflow across network

* Fix more double/float problem in serverpackethandler & remove implicit struct type for TileAnimationParams

* Fix connection unittests container

+++ src/client/clientevent.h
@@ -52,7 +52,7 @@ struct ClientEvent
 		//} none;
 		struct
 		{
-			u8 amount;
+			u16 amount;
 		} player_damage;
 		struct
 		{
